:recycle: REFACTOR: split abilities into multiple files
